Run `cargo fetch` before `cargo metadata` to get better output and errors
The `cargo metadata` wrapper suppresses stdout so that the in-progress messages for long lasting operatiions such as “Updating registry” are not shown. Also the error messages (e.g. when a dependency is not found) are wrapped in a cargo metadata error and are not colorized. By running `cargo fetch` first we fix both issues.