Add support for workspace subcrate building
This adds the configuration key `workspace-subcrate`, which lets the
root crate specify a subcrate within it's workspace to build as the
kernel executable.

For example, if your kernel's root crate is platform-agnostic, but you
have a subcrate that is for x86_64 (which would contain the kernel entry
point, etc).

Assuming your root crate is a library package, and your x86_64 subcrate
is a binary package, you can add a section like this to your root
crate's Cargo.toml:

    [package.metadata.bootimage]
    workspace-subcrate = "example-amd64"

In the `example-amd64` subcrate, you can then specify your `bootimage`
options normally, for example:

    [package.metadata.bootimage]
    default-target = "x86_64-example-os.json"

Once this is set up, running `bootimage run` (or any other `bootimage`
command using `bootimage::builder::Builder` such as `bootimage build`)
from the root of the crate (generally your repository root) will build
the subcrate as if you were in the subcrate's directory.

Note that this solution does not work if the root workspace manifest is
not a package itself, as the configuration entry allowing this is on
the `package.metadata.bootimage` table, and adding this to a non-package
workspace manifest causes cargo-metadata to say that there are missing
keys on the `package` table (defining a subtable of the `package` table
defines the `package` table, leading Cargo to believe that there should
be a package there).